Refactor: 본인 게시글 수정은 본인만

+++ 하은수/freelec-springboot2-webservice/src/main/java/com/jojoldu/book/springboot/service/posts/PostsService.java
@@ -38,10 +38,13 @@ public class PostsService {
     }
 
     @Transactional
-    public Long update(Long id, PostsUpdateRequestDto requestDto) {
-        Posts posts = postsRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("해당 게시글이 없습니다. id="+ id));
+    public Long update(Long id, PostsUpdateRequestDto requestDto, SessionUser sessionUser) {
+        Posts posts = postsRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("해당 게시글이 없습니다. id=" + id));
 
-        posts.update(requestDto.getTitle(), requestDto.getContent());
+        if (posts.getAuthor().getId().equals(sessionUser.getId())) {
+            posts.update(requestDto.getTitle(), requestDto.getContent());
+        }
 
         return id;
     }
@@ -58,9 +61,12 @@ public class PostsService {
     }
 
     @Transactional
-    public void delete (Long id) {
-        Posts posts = postsRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("해당 게시글이 없습니다. id="+ id));
+    public void delete(Long id, SessionUser sessionUser) {
+        Posts posts = postsRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("해당 게시글이 없습니다. id=" + id));
 
-        postsRepository.delete(posts);
+        if (posts.getAuthor().getId().equals(sessionUser.getId())) {
+            postsRepository.delete(posts);
+        }
     }
 }
